restructuring filtering module
