+++ python/drned_xmnr/op/filtering/__init__.py
@@ -1,0 +1,2 @@
+from .cort import filter_sink, drop
+from .filtering import *
